backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Azure OpenAI
openai==1.10.0
azure-identity==1.15.0

# Audio Processing
pydub==0.25.1
python-magic-bin==0.4.14  # For file type detection on Windows

# Environment Variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Requests
httpx==0.26.0
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
